# packages/shared/openapi.yaml
openapi: 3.0.3
info:
  title: Carmel Agent API
  version: 1.0.0
  description: Authoritative HTTP contract for native and web Carmel Agent clients. Streams are NDJSON, never SSE.
servers:
  - url: /
security:
  - carmelSession: []
...
paths:
...
  /api/agents/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    put: {operationId: upsertAgent, tags: [Agents], requestBody: {$ref: '#/components/requestBodies/Agent'}, responses: {'200': {description: Redacted agent, content: {application/json: {schema: {$ref: '#/components/schemas/AgentConfig'}}}}, '409': {$ref: '#/components/responses/Error'}}}
    delete: {operationId: deleteAgent, tags: [Agents], responses: {'200': {$ref: '#/components/responses/JSONSuccess'}, '409': {$ref: '#/components/responses/Error'}}}
  /api/agents/{id}/settings:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get: {operationId: getAgentSettings, tags: [Agents], responses: {'200': {description: Owner settings, content: {application/json: {schema: {$ref: '#/components/schemas/AgentConfig'}}}}}}
  /api/agents/{id}/commands:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get: {operationId: getAgentCommands, tags: [Agents], responses: {'200': {description: Commands, content: {application/json: {schema: {type: object, additionalProperties: true}}}}}}
...
components:
  securitySchemes:
    carmelSession: {type: apiKey, in: cookie, name: carmel_session}
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string}}
...
  responses:
    Error: {description: Error, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    JSONSuccess: {description: Success, content: {application/json: {schema: {$ref: '#/components/schemas/OK'}}}}
...
  requestBodies:
...
    Agent: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AgentRequest'}}}}
...
  schemas:
...
    Error: {type: object, required: [error], properties: {error: {type: string}}}
    OK: {type: object, required: [ok], properties: {ok: {type: boolean}}}
...
    ThinkingLevel: {type: string, enum: [off, minimal, low, medium, high, xhigh, max]}
...
    AgentRequest: {type: object, required: [shared, name, description, workingDirMode, workingDir, mounts, systemPrompt, promptTemplates, permissions, defaultModelRefId, defaultThinkingLevel], additionalProperties: false, properties: {shared: {type: boolean}, name: {type: string}, description: {type: string}, workingDirMode: {type: string, enum: [default, manual]}, workingDir: {type: string}, defaultWorkingDir: {type: string}, mounts: {type: array, items: {type: object, additionalProperties: true}}, systemPrompt: {type: string}, promptTemplates: {type: array, items: {type: object, additionalProperties: true}}, permissions: {type: object, additionalProperties: {type: boolean}}, defaultModelRefId: {type: string}, defaultThinkingLevel: {$ref: '#/components/schemas/ThinkingLevel'}}}
    AgentConfig: {type: object, required: [id, ownerUserId, shared, name, description, workingDirMode, workingDir, mounts, systemPrompt, promptTemplates, permissions, defaultModelRefId, defaultThinkingLevel, createdAt, updatedAt], additionalProperties: false, properties: {id: {type: string}, ownerUserId: {type: string}, shared: {type: boolean}, name: {type: string}, description: {type: string}, workingDirMode: {type: string, enum: [default, manual]}, workingDir: {type: string}, defaultWorkingDir: {type: string}, mounts: {type: array, items: {type: object, additionalProperties: true}}, systemPrompt: {type: string}, promptTemplates: {type: array, items: {type: object, additionalProperties: true}}, permissions: {type: object, additionalProperties: {type: boolean}}, defaultModelRefId: {type: string}, defaultThinkingLevel: {$ref: '#/components/schemas/ThinkingLevel'}, createdAt: {type: number}, updatedAt: {type: number}}}
...